# Refine CUDA atomicAdd for FP16 by CUDA primitive methods